feat: moved functions in utilities.ts into separate files in utilities folder
