+++ automation-tests/test/utilities/index.ts
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2023, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export * from './commandPrompt';
+export * from './fileSystem';
+export * from './miscellaneous';
+export * from './notifications';
+export * from './outputView';
+export * from './sideBar';
+export * from './statusBar';
+export * from './terminalView';
